requirements.txt
# Home Finder Dependencies

# Web framework
flask>=3.0.0
gunicorn>=21.0.0

# Scraping
rightmove-webscraper>=1.1
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
lxml>=5.0.0

# Data processing
pandas>=2.1.0
pyyaml>=6.0.0

# Geospatial
geopy>=2.4.0

# API clients
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
tenacity>=8.2.0  # Retry logic

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Dev
black>=23.0.0
ruff>=0.1.0
